requirements.txt
absl-py==2.1.0
aiohttp==3.8.1
aiosignal==1.2.0
alpaca-py==0.5.1
alpaca-trade-api==2.3.0
appdirs==1.4.4
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.1.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.2
blinker==1.6.2
cachetools==5.3.3
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.0.7
cramjam==2.6.2
cryptography==39.0.2
cycler==0.11.0
deprecation==2.1.0
dnspython==2.3.0
docopt==0.6.2
exchange-calendars==4.1.1
fastparquet==0.8.1
filelock==3.12.3
Flask==2.2.3
Flask-Cors==3.0.10
Flask-Login==0.6.2
flatbuffers==24.3.7
fonttools==4.38.0
frozendict==2.3.5
frozenlist==1.3.1
fsspec==2023.3.0
gast==0.4.0
google-auth==2.29.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
html5lib==1.1
idna==3.4
importlib-resources==5.12.0
ipywidgets==8.0.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
jupyterlab-widgets==3.0.5
keras==2.13.1
kiwisolver==1.4.4
korean-lunar-calendar==0.2.1
libclang==18.1.1
lxml==4.9.2
Markdown==3.6
MarkupSafe==2.1.3
matplotlib==3.7.1
mpmath==1.3.0
msgpack==1.0.3
multidict==6.0.2
multitasking==0.0.11
networkx==3.1
numpy==1.24.4
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==21.3
pandas
pandas-datareader==0.10.0
Pillow==9.4.0
pipreqs==0.4.11
plotext==5.0.2
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.9.1
pyluach==2.0.0
pymongo==4.3.3
PyNaCl==1.5.0
pyparsing==3.0.9
pyrh==2.0
python-dotenv==1.0.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.28.2
requests-oauthlib==2.0.0
rsa==4.9
schedule==1.1.0
scikit-learn==1.1.1
scipy==1.10.1
seaborn==0.11.2
sklearn==0.0
soupsieve==2.4
sympy==1.12
TA-Lib==0.4.25
termcolor==2.4.0
threadpoolctl==3.1.0
toolz==0.12.0
torch==2.0.1
tqdm==4.65.0
tzdata==2024.1
tzlocal==4.2
urllib3==1.26.14
webencodings==0.5.1
websocket-client==1.3.3
websockets==10.3
Werkzeug==2.3.4
widgetsnbextension==4.0.5
wrapt==1.16.0
xgboost==1.7.5
yarg==0.1.9
yarl==1.8.1
yfinance==0.2.12
zipp==3.15.0
pynacl==1.5.0
